% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnm.R
\name{gridpoints}
\alias{gridpoints}
\title{Grid points}
\usage{
gridpoints(x, beta, grid)
}
\arguments{
\item{x}{an object of a class for data.}

\item{beta}{instrumental parameter in a semiparametric mixture.}

\item{grid}{number of grid points to be generated.}
}
\value{
A numeric vector containing grid points.
}
\description{
A generic method used to return a vector of grid
  points used for searching local maxima of the gradient funcion.
}
\author{
Yong Wang <yongwang@auckland.ac.nz>
}
