% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnm.R
\name{initial}
\alias{initial}
\title{Initialization for a nonparametric/semiparametric mixture}
\usage{
initial(x, beta, mix, kmax)
}
\arguments{
\item{x}{an object of a class for data.}

\item{beta}{instrumental parameter in a semiparametric mixture.}

\item{mix}{an object of class \code{disc} for the mixing
distribution.}

\item{kmax}{the maximum allowed number of support points used.}
}
\value{
\item{beta}{an initialized value of beta}

\item{mix}{an initialised or updated object of class \code{disc}
for the mixing distribution.}
}
\description{
A generic method used to return an initialization for
  a nonparametric/semiparametric mixture.
}
\author{
Yong Wang <yongwang@auckland.ac.nz>
}
