% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnm.R
\name{initial0}
\alias{initial0}
\title{Initialisation}
\usage{
initial0(x, init = NULL, kmax = NULL)
}
\arguments{
\item{x}{an object of a class for data.}

\item{init}{a list with initial values for beta and mix (as in the
output of \code{initial})}

\item{kmax}{the maximum allowed number of support points used.}
}
\value{
\item{beta}{an initialized value of beta}

\item{mix}{an initialised or updated object of class \code{disc}
for the mixing distribution.}
}
\description{
The functions examines whether the initial values are
  proper. If not, proper ones are provided, by employing the
  function "initial" provided by the class.
}
\author{
Yong Wang <yongwang@auckland.ac.nz>
}
