% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnm.R
\name{llex}
\alias{llex}
\title{Log-likleihood Extra Term.}
\usage{
llex(x, beta, mix)
}
\arguments{
\item{x}{an object of a class for data.}

\item{beta}{instrumental parameter in a semiparametric mixture.}

\item{mix}{an object of class \code{disc} for the mixing
distribution.}
}
\value{
a scalar value
}
\description{
Value of possibly an extra term in the log-likleihood
  function for the instrumental parameter beta
}
\author{
Yong Wang <yongwang@auckland.ac.nz>
}
