% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnm.R
\name{llexdb}
\alias{llexdb}
\title{Derivative of the log-likleihood Extra Term}
\usage{
llexdb(x, beta, mix)
}
\arguments{
\item{x}{an object of a class for data.}

\item{beta}{instrumental parameter in a semiparametric mixture.}

\item{mix}{an object of class \code{disc} for the mixing
distribution.}
}
\value{
a scalar value
}
\description{
Derivative of the log-likleihood extra term wrt beta
}
\author{
Yong Wang <yongwang@auckland.ac.nz>
}
