% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/disc.R
\name{print.disc}
\alias{print.disc}
\title{Prints a discrete distribution function}
\usage{
\method{print}{disc}(x, ...)
}
\arguments{
\item{x}{an object of class \code{disc}.}

\item{...}{arguments passed on to function \code{print}.}
}
\description{
Class \code{disc} is used to represent an arbitrary
  univariate discrete distribution with a finite number of support
  points.

Function \code{disc} creates an object of class \code{disc}, given
 the support points and probability values at these points.

Function \code{print.disc} prints the discrete distribution.
}
\examples{

(d = disc(pt=c(0,4), pr=c(0.3,0.7)))
 
}
\seealso{
\code{\link{cnm}}, \code{\link{cnmms}}.
}
\author{
Yong Wang <yongwang@auckland.ac.nz>
}
