% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npnorm.R
\name{sort.npnorm}
\alias{sort.npnorm}
\title{Sorting of an Object of Class \code{npnorm}}
\usage{
\method{sort}{npnorm}(x, decreasing = FALSE, ...)
}
\arguments{
\item{x}{an object of class \code{npnorm}.}

\item{decreasing}{logical, in the decreasing (default) or
increasing order.}

\item{...}{arguments passed to function \code{order}.}
}
\description{
Function \code{sort.npnorm} sorts an object of class
   \code{npnorm} in the order of the obsersed values.
}
\examples{

mix = disc(pt=c(0,4), pr=c(0.3,0.7))  # a discrete distribution
x = rnpnorm(20, mix, sd=1)
sort(x)

 
}
