% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nppois.R
\name{sort.nppois}
\alias{sort.nppois}
\title{Sorting of an Object of Class \code{nppois}}
\usage{
\method{sort}{nppois}(x, decreasing = FALSE, ...)
}
\arguments{
\item{x}{an object of class \code{nppois}.}

\item{decreasing}{logical, in the decreasing (default) or
increasing order.}

\item{...}{arguments passed to function \code{order}.}
}
\description{
Function \code{sort.nppois} sorts an object of class
  \code{nppois} in the order of the obsersed values.
}
\examples{

mix = disc(pt=c(0,4), pr=c(0.3,0.7))  # a discrete distribution
x = rnppois(20, mix)
sort(x)

 
}
