% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnm.R
\name{suppspace}
\alias{suppspace}
\title{Support space}
\usage{
suppspace(x, beta)
}
\arguments{
\item{x}{an object of a class for data.}

\item{beta}{instrumental parameter in a semiparametric mixture.}
}
\value{
A vector of length 2.
}
\description{
Range of the mixing variable (theta).
}
\author{
Yong Wang <yongwang@auckland.ac.nz>
}
