% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnm.R
\name{valid}
\alias{valid}
\title{Valid parameter values}
\usage{
valid(x, beta, theta)
}
\arguments{
\item{x}{an object of a class for data.}

\item{beta}{instrumental parameter in a semiparametric mixture.}

\item{theta}{values of the mixing variable.}
}
\value{
A logical value.
}
\description{
A generic method used to return \code{TRUE} if the
  values of the paramters use for a nonparametric/semiparametric
  mixture are valid, or \code{FALSE} if otherwise.
}
\author{
Yong Wang <yongwang@auckland.ac.nz>
}
