% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnm.R
\name{weight}
\alias{weight}
\title{Weights}
\usage{
weight(x, beta)
}
\arguments{
\item{x}{an object of a class for data.}

\item{beta}{instrumental parameter in a semiparametric mixture.}
}
\value{
a numeric vector of the weights.
}
\description{
Weights or frequencis of observations.
}
\author{
Yong Wang <yongwang@auckland.ac.nz>
}
